+++ src/util/RandomNumberMaker.js
@@ -1,4 +1,4 @@
-const { LESS_THAN_TEN } = require("../Constants/Constants");
+const { LESS_THAN_TEN } = require("../constant/Constants");
 
 const makeRandomNumber = () => {
   return Math.floor(Math.random() * LESS_THAN_TEN); // 0 ~ 9 사이의 정수
