+++ src/util/SplitAndTrimString.js
@@ -1,4 +1,4 @@
-const { COMMA } = require("../Constants/Constants");
+const { COMMA } = require("../constant/Constants");
 
 const splitAndTrimString = (string) => {
   return string.split(COMMA).map((str) => str.trim());
